jakaLi0309numberone

+++ repo/jcart-administration-back/src/main/java/io/ljh/jcartadministrationback/po/ReturnHistory.java
@@ -15,7 +15,7 @@ public class ReturnHistory {
 
     private Boolean customerNotified;
 
-    public Long getReturnHistoryId() {
+    public Integer getReturnHistoryId() {
         return returnHistoryId;
     }
